refactor: Organized Codebase
