+++ src/app/page.jsx
@@ -1,12 +1,12 @@
 import React from "react";
-import CustomerReview from "./about/review/page";
 import HeroSection from "@/components/Landing/HeroSection";
 import Features from "@/components/Landing/Features";
 import UseCases from "@/components/Landing/UseCases";
 import Integrations from "@/components/Landing/Integrations";
 import CompetitiveEdge from "@/components/Landing/CompetitiveEdge";
-import Faq from "@/components/home/Faq";
-import Analytics from "@/components/home/Analytics";
+import Faq from "@/components/Landing/Faq";
+import Analytics from "@/components/Landing/Analytics";
+import CustomerSlider from "@/components/Landing/CustomerSlider";
 
 export default function HomePage() {
   return (
@@ -17,8 +17,8 @@ export default function HomePage() {
       <Integrations />
       <CompetitiveEdge />
       <Analytics />
-      <Faq/>
-      <CustomerReview />
+      <Faq />
+      <CustomerSlider />
     </div>
   );
 }
